Fix cache source test runner for src-prec and invokedyn insn
As is defining the cache src-prec option in soot
defines it as the default if src-prec is not specified.
This will need to be fixed, but for now the test runner
portion that creates the baseline jimple file should specify
class as the preferred. Also this commit adds the identification
of the invokedynamic test to the test runner.

Signed-off-by: knewbury <knewbury@ualberta.ca>