Update pom for compiler and asm vs
Finally commit change for soot pom to build with
the OpenJ9 javac and provide the cp for that.
Also revert the soot usage of asm from 7.1 in a recent
develop branch merge, back to 5.2 which is what most of the
testing assumed. It may have been fine to leave but we can come
back to this later.

Signed-off-by: knewbury <knewbury@ualberta.ca>